订单系统修改

+++ src/main/java/ordersSystem/trans/QueryWaitPurchaseOrdersInfo.java
@@ -3,24 +3,29 @@ package ordersSystem.trans;
 public class QueryWaitPurchaseOrdersInfo extends QueryOrdersInfo{
 	private String goodPriority;
 	private Double goodWeight;
-	private Integer ordersPrice;
+	private Double ordersPrice;
+
 	public String getGoodPriority() {
 		return goodPriority;
 	}
+
 	public void setGoodPriority(String goodPriority) {
 		this.goodPriority = goodPriority;
 	}
+
 	public Double getGoodWeight() {
 		return goodWeight;
 	}
+
 	public void setGoodWeight(Double goodWeight) {
 		this.goodWeight = goodWeight;
 	}
-	public Integer getOrdersPrice() {
+
+	public Double getOrdersPrice() {
 		return ordersPrice;
 	}
-	public void setOrdersPrice(Integer ordersPrice) {
+
+	public void setOrdersPrice(Double ordersPrice) {
 		this.ordersPrice = ordersPrice;
 	}
-
 }
